refactor: modify blacklist modal to be used for add edit and delete
